requirements.txt
ultralytics 
opencv-python
numpy
pynput
maturin
supervision
pyyaml
pathlib